+++ client/src/components/Body.jsx
@@ -1,11 +1,20 @@
+import React from 'react';
 import Container from 'react-bootstrap/Container';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Card from 'react-bootstrap/Card';
 import Button from 'react-bootstrap/esm/Button';
+import { useNavigate } from 'react-router';
 
 const Body = (props) => {
+
   const {setUserName} = props
+  const nav = useNavigate()
+
+  const selectUser = (userType) => {
+    setUserName(userType)
+    nav(`/dashboard`)
+  }
 
   return (
    <Container className='p-2'>
@@ -15,11 +24,10 @@ const Body = (props) => {
             {/* <Card.Img variant="top" src="holder.js/100px180" /> */}
             <Card.Body>
               <Card.Title>Login As Job Poster</Card.Title>
-              {/* <Card.Text>
-                Some quick example text to build on the card title and make up the
-                bulk of the card's content.
-              </Card.Text> */}
-              <Button onClick={() => setUserName('Job Poster')} variant="primary">Go somewhere</Button>
+              <Card.Text>
+                View Hero for Hire with the functionalities provided a non-superhero user
+              </Card.Text>
+              <Button onClick={() => selectUser('Job Poster')} variant="primary">Go somewhere</Button>
             </Card.Body>
           </Card>
          </Col>
@@ -28,11 +36,10 @@ const Body = (props) => {
             {/* <Card.Img variant="top" src="holder.js/100px180" /> */}
             <Card.Body>
               <Card.Title>Login As Superhero</Card.Title>
-              {/* <Card.Text>
-                Some quick example text to build on the card title and make up the
-                bulk of the card's content.
-              </Card.Text> */}
-              <Button onClick={() => setUserName('Superhero')}  variant="primary">Go somewhere</Button>
+              <Card.Text>
+                View Hero for Hire with the functionalities provided a non-superhero user
+              </Card.Text>
+              <Button onClick={() => selectUser('Superhero')}  variant="primary">Go somewhere</Button>
             </Card.Body>
           </Card>
          </Col>
