+++ client/src/components/DisplayApplicantProfile.jsx
@@ -9,9 +9,23 @@ import UserProfile from './UserProfile';
 const DisplayApplicantProfile = (props) => {
    const {currentId} = props
    const [isUserFave, setIsUserFave] = useState(null)
-   // const [applicantId, setApplicantId] = useState(null)
+   const [displayedUserId, setDisplayedUserId] = useState(useParams().id)
    const nav = useNavigate()
-   const applicantId = useParams().id
+   
+   useEffect(() => {
+      if (isUserFave == null) {
+         if (currentId && displayedUserId) fetchIsFave()
+      } else setIsUserFave(null)
+   },[currentId, displayedUserId])
+
+   const fetchIsFave = () => {
+      const url = `http://localhost:5000/api/users/${currentId}`
+         fetch(url)
+         .then(res=>res.json()).then(data => setIsUserFave(() => data.favourites.includes(displayedUserId)))
+   }
+
+   // const displayedUserId = useParams().id
+   
    return (
       <>
          <Container fluid className='p-2 return-outer d-flex justify-content-center'>
@@ -24,7 +38,7 @@ const DisplayApplicantProfile = (props) => {
          <Container className='p-2'>
             <Row className='p-2'>
                <Col xs={12}>
-                  <UserProfile userId={applicantId} isFave={true}/>
+                  <UserProfile userId={displayedUserId} currentId={currentId} isUserFave={isUserFave} setDisplayedUserId={setDisplayedUserId} />
                </Col>
             </Row>
          </Container>
